update(data): rename pojo to request
