+++ stocky-api/src/main/java/com/jamesaworo/stocky/features/product/data/request/ProductSearchRequest.java
@@ -1,4 +1,4 @@
-package com.jamesaworo.stocky.features.product.data.pojo;
+package com.jamesaworo.stocky.features.product.data.request;
 
 /**
  * @author Aworo James
